+++ src/core.jl
@@ -1,5 +1,5 @@
-
 include("utils.jl")
+include("RBMBase.jl")
 include("rbm.jl")
 include("nets.jl")
 include("io.jl")
